Combine release summary tables into a single "Proteome Sources and Statistics" table
