--- Site/webapp/wdkCustomization/js/client/controllers/GenomeSourcesController.scss
@@ -1,3 +1,0 @@
-.ReleaseSummary.GenomeSources .EmptyDescription {
-  color: gray;
-}
