+++ Site/webapp/wdkCustomization/js/client/controllers/ProteomeSummaryController.scss
@@ -1,0 +1,12 @@
+.ReleaseSummary.ProteomeSummary {
+  .SearchBoxContainer {
+    display: grid;
+    grid-template-columns: auto auto;
+    align-items: center;
+    gap: 0.25em;
+  }
+
+  .EmptyDescription {
+    color: gray;
+  }
+}
