+++ Site/webapp/wdkCustomization/js/client/routes.tsx
@@ -4,9 +4,8 @@ import { RouteComponentProps } from 'react-router';
 
 import { RouteEntry } from 'wdk-client/Core/RouteEntry';
 
-import { GenomeSourcesController } from 'ortho-client/controllers/GenomeSourcesController';
-import { GenomeStatisticsController } from 'ortho-client/controllers/GenomeStatisticsController';
 import { OrthoMCLHomePageController } from 'ortho-client/controllers/OrthoMCLHomePageController';
+import { ProteomeSummaryController } from 'ortho-client/controllers/ProteomeSummaryController';
 import { GroupClusterGraphController } from 'ortho-client/controllers/GroupClusterGraphController';
 
 export function wrapRoutes(ebrcRoutes: RouteEntry[]): RouteEntry[] {
@@ -27,12 +26,8 @@ export function wrapRoutes(ebrcRoutes: RouteEntry[]): RouteEntry[] {
       }
     },
     {
-      path: '/genome-statistics',
-      component: GenomeStatisticsController
-    },
-    {
-      path: '/genome-sources',
-      component: GenomeSourcesController,
+      path: '/release-summary',
+      component: ProteomeSummaryController
     },
     ...ebrcRoutes
   ];
